Extract declareVictor function to its own file
