+++ actions/index.js
@@ -2,6 +2,7 @@ const ping = require('./ping')
 const startPoll = require('./startpoll')
 const submit = require('./submit')
 const closePoll = require('./closepoll')
+const declareVictor = require('./declarevictor')
 const help = require('./help')
 
 module.exports = {
@@ -9,5 +10,6 @@ module.exports = {
     startPoll,
     submit,
     closePoll,
+    declareVictor,
     help,
 }
